SVN r1798 (odaraven)

+++ client/src/r_plane.h
@@ -4,6 +4,7 @@
 // $Id$
 //
 // Copyright (C) 1993-1996 by id Software, Inc.
+// Copyright (C) 2006-2010 by The Odamex Team.
 //
 // This program is free software; you can redistribute it and/or
 // modify it under the terms of the GNU General Public License
@@ -31,7 +32,7 @@
 #define PL_SKYFLAT (0x80000000)
 
 // Visplane related.
-extern	short*			lastopening;
+extern	int*			lastopening;
 
 
 typedef void (*planefunction_t) (int top, int bottom);
@@ -39,8 +40,8 @@ typedef void (*planefunction_t) (int top, int bottom);
 extern planefunction_t	floorfunc;
 extern planefunction_t	ceilingfunc_t;
 
-extern short			*floorclip;
-extern short			*ceilingclip;
+extern int  			*floorclip;
+extern int  			*ceilingclip;
 
 extern fixed_t			*yslope;
 extern fixed_t			*distscale;
